fix: Resolve build errors in IntelligentQuerySynthesizer
- Corrects all `Cannot find module` errors by replacing incorrect relative import paths with the project's standard `@/*` aliases.
- Resolves the complex TypeScript error `Type 'string' is not assignable to type 'SchemaType.OBJECT'` by applying an `:any` type annotation to the schema constants as a workaround for the SDK's strict typings.
- Updates the `generateContent` call to conform to the current version of the `@google/generative-ai` SDK.
